♻️ Extension 관련 파일 정리
