--- JipJung/JipJung/UI/Common/SystemConstants.swift
@@ -1,16 +1,0 @@
-//
-//  SystemConstants.swift
-//  JipJung
-//
-//  Created by Soohyeon Lee on 2021/11/08.
-//
-
-import UIKit
-
-enum SystemConstants {
-    static let deviceScreenSize: (width: CGFloat, height: CGFloat) = (
-        width: UIScreen.main.bounds.width,
-        height: UIScreen.main.bounds.height
-    )
-    static let statusBarHeight: CGFloat = 47
-}
