+++ JipJung/JipJung/UI/Common/UIView+Extension.swift
@@ -11,4 +11,13 @@ extension UIView {
     func makeCircle() {
         layer.cornerRadius = bounds.height / 2
     }
+    
+    func makeBlurBackground(style: UIBlurEffect.Style = .light) {
+        let blurEffect = UIBlurEffect(style: style)
+        let blurView = UIVisualEffectView(effect: blurEffect)
+        insertSubview(blurView, at: 0)
+        blurView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
 }
